Lib testing refactor
